Great Adventures
Far over the misty mountains cold
To dungeons deep and caverns old
We must away ere break of day
To find our long-forgotten gold.

The pines were roaring on the height,
The winds were moaning in the night.
The fire was red, it flaming spread;
The trees like torches blazed with light.